refactor(edge-jobs): move tasks handler to edgejobs
