+++ api/http/handler/edgejobs/edgejob_tasklogs_collect.go
@@ -1,4 +1,4 @@
-package edgejobtasks
+package edgejobs
 
 import (
 	"net/http"
